+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+# UNRELEASED
+
+## Added
+- `Events.add_all` to set data that should be included with all requests.
+
 # 0.1.0 - 2018-07-09
 
 First release!